requirements.txt
requests[socks]
feedparser
python-dotenv
jinja2
sumy
nltk
numpy
trafilatura
youtube-transcript-api
tavily-python
googlenewsdecoder
flask
groq
